Fixed applauncher compile
<==== Mercurial data ====>
Branch: master

+++ applauncher/src/named_pipe_socket_unix.h
@@ -10,7 +10,7 @@
 class NamedPipeSocketImpl
 {
 public:
-    static const DWORD BUFSIZE = 4*1024;
+    static const size_t BUFSIZE = 4*1024;
 
     int hPipe;
